+++ Types/CommandAST/get_IsPiped.ps1
@@ -1,0 +1,2 @@
+($this.Parent -is [Management.Automation.Language.PipelineAst]) -and
+($this.Parent.PipelineElements.Count -gt 1)